refactor(update): split installer-binary responsibilities
